Fix linter findings (#178)
* [linter] Fix unnecessary conversion (unconvert)

* [linter] Fix fmt.Sprintf can be replaced with faster strconv.FormatInt (perfsprint)

+++ x/dao/keeper/distribution.go
@@ -106,7 +106,7 @@ func (k Keeper) GetDistributionForReissuedTokens(ctx sdk.Context, blockHeight in
 	var overallAmount uint64
 	for index, obj := range reissuances {
 		if (index == 0 && lastPoP == 0 && obj.BlockHeight == 0) || //corner case (beginning of he chain)
-			(int64(lastPoP) < int64(obj.BlockHeight) && int64(obj.BlockHeight) <= blockHeight) {
+			(lastPoP < obj.BlockHeight && obj.BlockHeight <= blockHeight) {
 			amount, err := getUint64FromTxString(ctx, obj.Rawtx)
 			if err == nil {
 				overallAmount = overallAmount + amount
